requirements.txt
torch
torchvision
albumentations
numpy
pillow
tqdmgit add .